requirements-minimal.txt
# 最小化依賴版本 - 移除版本限制讓 pip 自動解析
line-bot-sdk
flask
google-cloud-aiplatform
cloudinary
redis
gunicorn